box-installer:  change trigger to init extendRoot
Newer openwrt releases activate all available init scripts at boot time.
The check if the init script was enabled does not work anymore for a
valid trigger.
extendRoot does create missing mount point folder, so these checks do not
work as well.

We need to rely on known UCI changes. If these are available, we know
that the package was initialized or not.